struct: Move trees to separate package(dir)
