+++ binaryTree/src/main/kotlin/org/tree/binaryTree/trees/RBTree.kt
@@ -1,4 +1,7 @@
-package org.tree.binaryTree
+package org.tree.binaryTree.trees
+
+import org.tree.binaryTree.RBNode
+import org.tree.binaryTree.TemplateBalanceBSTree
 
 class RBTree<T : Comparable<T>> : TemplateBalanceBSTree<T, RBNode<T>>() {
     private fun findParentForNewNode(curNode: RBNode<T>?, obj: T): RBNode<T>? {
